refactor: BookCart의 로직을 템플릿으로 이동
